fix post merge problems
