+++ src/cairo_types/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod builtins;
+pub(crate) mod structs;
 pub(crate) mod traits;
 pub(crate) mod trie;
