+++ src/hints/mod.rs
@@ -1,15 +1,6 @@
-pub mod block_context;
-pub mod builtins;
-pub mod execution;
-pub mod state;
-pub mod syscalls;
-#[cfg(test)]
-mod tests;
-mod unimplemented;
-mod vars;
-
 use std::collections::{HashMap, HashSet};
 
+use cairo_vm::Felt252;
 use cairo_vm::hint_processor::builtin_hint_processor::builtin_hint_processor_definition::{
     BuiltinHintProcessor, HintProcessorData,
 };
@@ -23,13 +14,21 @@ use cairo_vm::types::relocatable::MaybeRelocatable;
 use cairo_vm::vm::errors::hint_errors::HintError;
 use cairo_vm::vm::runners::cairo_runner::{ResourceTracker, RunResources};
 use cairo_vm::vm::vm_core::VirtualMachine;
-use cairo_vm::Felt252;
 use indoc::indoc;
 use num_bigint::BigInt;
 
 use crate::execution::helper::ExecutionHelperWrapper;
-use crate::config::DEFAULT_INPUT_PATH;
 use crate::io::input::StarknetOsInput;
+
+pub mod block_context;
+pub mod builtins;
+pub mod execution;
+pub mod state;
+pub mod syscalls;
+#[cfg(test)]
+mod tests;
+mod unimplemented;
+mod vars;
 
 type HintImpl = fn(
     &mut VirtualMachine,
@@ -39,14 +38,12 @@ type HintImpl = fn(
     &HashMap<String, Felt252>,
 ) -> Result<(), HintError>;
 
-static HINTS: [(&str, HintImpl); 78] = [
-    // (BREAKPOINT, breakpoint),
-    (STARKNET_OS_INPUT, starknet_os_input),
-    (INITIALIZE_STATE_CHANGES, initialize_state_changes),
+static HINTS: [(&str, HintImpl); 77] = [
     (INITIALIZE_CLASS_HASHES, initialize_class_hashes),
     (INITIALIZE_STATE_CHANGES, initialize_state_changes),
     (IS_N_GE_TWO, is_n_ge_two),
     (IS_ON_CURVE, is_on_curve),
+    (OS_INPUT_TRANSACTIONS, os_input_transactions),
     (SEGMENTS_ADD, segments_add),
     (SEGMENTS_ADD_TEMP, segments_add_temp),
     (SET_AP_TO_ACTUAL_FEE, set_ap_to_actual_fee),
@@ -54,7 +51,6 @@ static HINTS: [(&str, HintImpl); 78] = [
     (SKIP_TX, skip_tx),
     (STARKNET_OS_INPUT, starknet_os_input),
     (START_TX, start_tx),
-    (TRANSACTIONS_LEN, transactions_len),
     (block_context::BLOCK_NUMBER, block_context::block_number),
     (block_context::BLOCK_TIMESTAMP, block_context::block_timestamp),
     (block_context::CHAIN_ID, block_context::chain_id),
@@ -73,20 +69,18 @@ static HINTS: [(&str, HintImpl); 78] = [
     (execution::CHECK_IS_DEPRECATED, execution::check_is_deprecated),
     (execution::CONTRACT_ADDRESS, execution::contract_address),
     (execution::END_TX, execution::end_tx),
-    (execution::END_TX, execution::end_tx),
     (execution::ENTER_CALL, execution::enter_call),
-    (execution::ENTER_CALL, execution::enter_call),
-    (execution::ENTER_SCOPE_SYSCALL_HANDLER, execution::enter_scope_syscall_handler),
     (execution::ENTER_SCOPE_SYSCALL_HANDLER, execution::enter_scope_syscall_handler),
     (execution::ENTER_SYSCALL_SCOPES, execution::enter_syscall_scopes),
     (execution::EXIT_CALL, execution::exit_call),
-    (execution::EXIT_CALL, execution::exit_call),
+    (execution::GEN_SIGNATURE_ARG, execution::gen_signature_arg),
     (execution::GET_STATE_ENTRY, execution::get_state_entry),
     (execution::IS_DEPRECATED, execution::is_deprecated),
     (execution::LOAD_NEXT_TX, execution::load_next_tx),
     (execution::OS_CONTEXT_SEGMENTS, execution::os_context_segments),
     (execution::PREPARE_CONSTRUCTOR_EXECUTION, execution::prepare_constructor_execution),
     (execution::RESOURCE_BOUNDS, execution::resource_bounds),
+    (execution::START_TX, execution::start_tx),
     (execution::TRANSACTION_VERSION, execution::transaction_version),
     (execution::TX_ACCOUNT_DEPLOYMENT_DATA, execution::tx_account_deployment_data),
     (execution::TX_ACCOUNT_DEPLOYMENT_DATA_LEN, execution::tx_account_deployment_data_len),
@@ -101,15 +95,10 @@ static HINTS: [(&str, HintImpl); 78] = [
     (execution::TX_PAYMASTER_DATA_LEN, execution::tx_paymaster_data_len),
     (execution::TX_RESOURCE_BOUNDS_LEN, execution::tx_resource_bounds_len),
     (execution::TX_TIP, execution::tx_tip),
-    (state::LOAD_EDGE, state::prepare_preimage_validation),
+    (state::LOAD_EDGE, state::load_edge),
     (state::SET_PREIMAGE_FOR_CLASS_COMMITMENTS, state::set_preimage_for_class_commitments),
     (state::SET_PREIMAGE_FOR_CURRENT_COMMITMENT_INFO, state::set_preimage_for_current_commitment_info),
     (state::SET_PREIMAGE_FOR_STATE_COMMITMENTS, state::set_preimage_for_state_commitments),
-    (execution::GEN_SIGNATURE_ARG, execution::gen_signature_arg),
-    (state::SET_PREIMAGE_FOR_CLASS_COMMITMENTS, state::set_preimage_for_class_commitments),
-    (state::SET_PREIMAGE_FOR_CURRENT_COMMITMENT_INFO, state::set_preimage_for_current_commitment_info),
-    (state::SET_PREIMAGE_FOR_STATE_COMMITMENTS, state::set_preimage_for_state_commitments),
-    (state::PREPARE_PREIMAGE_VALIDATION, state::prepare_preimage_validation),
     (syscalls::CALL_CONTRACT, syscalls::call_contract),
     (syscalls::DELEGATE_CALL, syscalls::delegate_call),
     (syscalls::DELEGATE_L1_HANDLER, syscalls::delegate_l1_handler),
@@ -127,15 +116,6 @@ static HINTS: [(&str, HintImpl); 78] = [
     (syscalls::SEND_MESSAGE_TO_L1, syscalls::send_message_to_l1),
     (syscalls::STORAGE_READ, syscalls::storage_read),
     (syscalls::STORAGE_WRITE, syscalls::storage_write),
-    // (execution::START_DEPLOY_TX, execution::start_deploy_tx),
-    // add missing execution hints
-    (SET_AP_TO_ACTUAL_FEE, set_ap_to_actual_fee),
-    (IS_ON_CURVE, is_on_curve),
-    (IS_N_GE_TWO, is_n_ge_two),
-    (START_TX, start_tx),
-    (SKIP_TX, skip_tx),
-    (SKIP_CALL, skip_call),
-    (OS_INPUT_TRANSACTIONS, os_input_transactions),
 ];
 
 /// Hint Extensions extend the current map of hints used by the VM.
@@ -484,111 +464,6 @@ pub fn skip_call(
     _constants: &HashMap<String, Felt252>,
 ) -> Result<(), HintError> {
     let mut execution_helper = exec_scopes.get::<ExecutionHelperWrapper>(vars::scopes::EXECUTION_HELPER)?;
-    execution_helper.skip_call();
-
-    Ok(())
-}
-
-pub const IS_N_GE_TWO: &str = "memory[ap] = to_felt_or_relocatable(ids.n >= 2)";
-
-pub fn is_n_ge_two(
-    vm: &mut VirtualMachine,
-    _exec_scopes: &mut ExecutionScopes,
-    ids_data: &HashMap<String, HintReference>,
-    ap_tracking: &ApTracking,
-    _constants: &HashMap<String, Felt252>,
-) -> Result<(), HintError> {
-    let n = get_integer_from_var_name(vars::ids::N, vm, ids_data, ap_tracking)?.into_owned();
-    let value = if n >= Felt252::TWO { Felt252::ONE } else { Felt252::ZERO };
-    insert_value_into_ap(vm, value)?;
-    Ok(())
-}
-
-pub const SET_AP_TO_ACTUAL_FEE: &str =
-    "memory[ap] = to_felt_or_relocatable(execution_helper.tx_execution_info.actual_fee)";
-
-pub fn set_ap_to_actual_fee(
-    vm: &mut VirtualMachine,
-    exec_scopes: &mut ExecutionScopes,
-    _ids_data: &HashMap<String, HintReference>,
-    _ap_tracking: &ApTracking,
-    _constants: &HashMap<String, Felt252>,
-) -> Result<(), HintError> {
-    let execution_helper = exec_scopes.get::<ExecutionHelperWrapper>(vars::scopes::EXECUTION_HELPER)?;
-    let actual_fee = execution_helper
-        .execution_helper
-        .borrow()
-        .tx_execution_info
-        .as_ref()
-        .ok_or(HintError::CustomHint("ExecutionHelper should have tx_execution_info".to_owned().into_boxed_str()))?
-        .actual_fee;
-
-    insert_value_into_ap(vm, Felt252::from(actual_fee.0))
-}
-
-pub const IS_ON_CURVE: &str = "ids.is_on_curve = (y * y) % SECP_P == y_square_int";
-
-pub fn is_on_curve(
-    vm: &mut VirtualMachine,
-    exec_scopes: &mut ExecutionScopes,
-    ids_data: &HashMap<String, HintReference>,
-    ap_tracking: &ApTracking,
-    _constants: &HashMap<String, Felt252>,
-) -> Result<(), HintError> {
-    let y: BigInt = exec_scopes.get(vars::ids::Y)?;
-    let y_square_int: BigInt = exec_scopes.get(vars::ids::Y_SQUARE_INT)?;
-    let sec_p: BigInt = exec_scopes.get(vars::ids::SECP_P)?;
-
-    let is_on_curve = (y.clone() * y) % sec_p == y_square_int;
-    let is_on_curve: Felt252 = if is_on_curve { Felt252::ONE } else { Felt252::ZERO };
-    insert_value_from_var_name(vars::ids::IS_ON_CURVE, is_on_curve, vm, ids_data, ap_tracking)?;
-
-    Ok(())
-}
-
-const START_TX: &str = "execution_helper.start_tx(tx_info_ptr=ids.deprecated_tx_info.address_)";
-
-pub fn start_tx(
-    vm: &mut VirtualMachine,
-    exec_scopes: &mut ExecutionScopes,
-    ids_data: &HashMap<String, HintReference>,
-    ap_tracking: &ApTracking,
-    _constants: &HashMap<String, Felt252>,
-) -> Result<(), HintError> {
-    let deprecated_tx_info_ptr =
-        get_relocatable_from_var_name(vars::ids::DEPRECATED_TX_INFO, vm, ids_data, ap_tracking)?;
-
-    let execution_helper = exec_scopes.get::<ExecutionHelperWrapper>(vars::scopes::EXECUTION_HELPER).unwrap();
-    execution_helper.start_tx(Some(deprecated_tx_info_ptr));
-
-    Ok(())
-}
-
-const SKIP_TX: &str = "execution_helper.skip_tx()";
-
-pub fn skip_tx(
-    _vm: &mut VirtualMachine,
-    exec_scopes: &mut ExecutionScopes,
-    _ids_data: &HashMap<String, HintReference>,
-    _ap_tracking: &ApTracking,
-    _constants: &HashMap<String, Felt252>,
-) -> Result<(), HintError> {
-    let execution_helper = exec_scopes.get::<ExecutionHelperWrapper>(vars::scopes::EXECUTION_HELPER).unwrap();
-    execution_helper.skip_tx();
-
-    Ok(())
-}
-
-const SKIP_CALL: &str = "execution_helper.skip_call()";
-
-pub fn skip_call(
-    _vm: &mut VirtualMachine,
-    exec_scopes: &mut ExecutionScopes,
-    _ids_data: &HashMap<String, HintReference>,
-    _ap_tracking: &ApTracking,
-    _constants: &HashMap<String, Felt252>,
-) -> Result<(), HintError> {
-    let mut execution_helper = exec_scopes.get::<ExecutionHelperWrapper>(vars::scopes::EXECUTION_HELPER).unwrap();
     execution_helper.skip_call();
 
     Ok(())
